+++ src/app/components/content-clusters-certificate.js
@@ -17,14 +17,7 @@ import GlobalActions from 'onedata-gui-common/mixins/components/global-actions';
 import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
 import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 import Looper from 'onedata-gui-common/utils/looper';
-import config from 'ember-get-config';
 import computedPipe from 'onedata-gui-common/utils/ember/computed-pipe';
-
-const {
-  time: {
-    reloadDelayForCertificateChange,
-  },
-} = config;
 
 /**
  * @type {Number} interval in ms
@@ -34,8 +27,6 @@ const webCertPollInterval = 2000;
 export default Component.extend(I18n, GlobalActions, {
   webCertManager: service(),
   onepanelServer: service(),
-  deploymentManager: service(),
-  providerManager: service(),
   globalNotify: service(),
   guiUtils: service(),
 
@@ -69,10 +60,6 @@ export default Component.extend(I18n, GlobalActions, {
    */
   showRedirectPage: false,
 
-  isEmergencyOnepanel: reads('onepanelServer.isEmergency'),
-
-  onepanelServiceType: reads('guiUtils.serviceType'),
-
   /**
    * True, if regenerate action is pending
    * @type {Ember.ComputedProperty<boolean>}
@@ -90,12 +77,6 @@ export default Component.extend(I18n, GlobalActions, {
    * @type {boolean}
    */
   refreshCert: true,
-
-  /**
-   * Using intermediate var for testing purposes
-   * @type {Location}
-   */
-  _location: location,
 
   /**
    * @type {Ember.ComputedProperty<boolean>}
@@ -161,10 +142,8 @@ export default Component.extend(I18n, GlobalActions, {
 
   schedulePageReload() {
     safeExec(this, 'set', 'showRedirectPage', true);
-    const _location = this.get('_location');
-    setTimeout(() => {
-      _location.reload();
-    }, reloadDelayForCertificateChange);
+    return this.get('webCertManager').reloadPageAfterWebCertChange()
+      .catch(() => safeExec(this, 'set', 'showRedirectPage', false));
   },
 
   actions: {