+++ src/app/services/web-cert-manager.js
@@ -9,10 +9,38 @@
 
 import { default as Service, inject as service } from '@ember/service';
 import { computed, get } from '@ember/object';
+import { reads } from '@ember/object/computed';
 import createDataProxyMixin from 'onedata-gui-common/utils/create-data-proxy-mixin';
+import { reject, resolve } from 'rsvp';
+import config from 'ember-get-config';
+import changeDomain from 'onepanel-gui/utils/change-domain';
+
+const {
+  time: {
+    reloadDelayForCertificateChange,
+  },
+} = config;
 
 export default Service.extend(createDataProxyMixin('webCert'), {
   onepanelServer: service(),
+  guiUtils: service(),
+  deploymentManager: service(),
+  providerManager: service(),
+
+  /**
+   * @type {Location}
+   */
+  _location: location,
+
+  /**
+   * @type {ComputedProperty<String>}
+   */
+  onepanelServiceType: reads('guiUtils.serviceType'),
+
+  /**
+   * @type {ComputedProperty<Boolean>}
+   */
+  isEmergencyOnepanel: reads('onepanelServer.isEmergency'),
 
   /**
    * @type {Ember.ComputedProperty<boolean>}
@@ -41,5 +69,42 @@ export default Service.extend(createDataProxyMixin('webCert'), {
         letsEncrypt,
       }
     );
+  },
+
+  /**
+   * @returns {Promise}
+   */
+  reloadPageAfterWebCertChange() {
+    return this.getDomainForRedirectAfterWebCertChange()
+      .then(domain => changeDomain(domain, {
+        location: this.get('_location'),
+        delay: reloadDelayForCertificateChange,
+      }));
+  },
+
+  /**
+   * @returns {Promise<String>}
+   */
+  getDomainForRedirectAfterWebCertChange() {
+    const {
+      onepanelServiceType,
+      isEmergencyOnepanel,
+      _location,
+    } = this.getProperties('onepanelServiceType', 'isEmergencyOnepanel', '_location');
+
+    if (isEmergencyOnepanel || onepanelServiceType === 'onezone') {
+      switch (onepanelServiceType) {
+        case 'oneprovider':
+          return this.get('providerManager').getProviderDetailsProxy()
+            .then(provider => provider && get(provider, 'domain'));
+        case 'onezone':
+          return this.get('deploymentManager').getClusterConfiguration()
+            .then(({ data: cluster }) => cluster && get(cluster, 'onezone.domainName'));
+        default:
+          return reject(`Invalid onepanelServiceType: ${onepanelServiceType}`);
+      }
+    } else {
+      return resolve(_location.hostname);
+    }
   },
 });